requirements.txt
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
drf-yasg==1.21.7
